% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data.example}
\alias{data.example}
\title{Simulated data for the example.}
\format{
A data list with 5 elements:
\describe{
\item{t.0}{time at which the surrogate is measured}
\item{t}{time at which the primary outcome is measured}
\item{xob}{observed survival time}
\item{s.ob}{surrogate information at t.0}
\item{deltaob}{event indicator}
\item{aob}{treatment indicator}
}
}
\usage{
data.example
}
\description{
Simulated data for the example.
}
\keyword{datasets}
