% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pte.R
\name{pte.survival}
\alias{pte.survival}
\title{Main estimation function}
\usage{
pte.survival(xob, s.ob, deltaob, aob, t, t.0, varind = 0, re = 100)
}
\arguments{
\item{re}{}
}
\value{

}
\description{
Main estimation function
}
